Create Photo component and pass errors back up to App
